3.1. Загрузка файлов на github [#243584]

+++ src/ru/job4j/calculate/Fit.java
@@ -2,11 +2,11 @@ package ru.job4j.calculate;
 
 public class Fit {
     public static double manWeight(short height) {
-        return ((double) height - 100) * 1.15;
+        return (height - 100) * 1.15;
     }
 
     public static double womanWeight(short height) {
-        return ((double) height - 110) * 1.15;
+        return (height - 110) * 1.15;
     }
 
     public static void main(String[] args) {
